+++ src/Entity/User/ProductOwner.php
@@ -3,10 +3,6 @@
 namespace App\Entity\User;
 
 use Doctrine\ORM\Mapping as ORM;
-use Doctrine\Common\Collections\Collection;
-use Doctrine\Common\Collections\ArrayCollection;
-
-use App\Entity\Project\Project;
 
 /**
  * @ORM\Entity()
@@ -14,31 +10,8 @@ use App\Entity\Project\Project;
  */
 class ProductOwner extends User
 {
-    /**
-     * @ORM\OneToMany(targetEntity="App\Entity\Project\Project", mappedBy="productOwner")
-     */
-    protected $projects;
-    
-    public function __construct()
-    {
-        parent::__construct();
-        $this->projects = new ArrayCollection();
-    }
-    
     public function getType(): string
     {
         return self::TYPE_PRODUCT_OWNER;
-    }
-    
-    public function addProject(Project $project): ProductOwner
-    {
-        $this->projects->add($project);
-        
-        return $this;
-    }
-    
-    public function getProjects(): Collection
-    {
-        return $this->projects;
     }
 }